setup stacks and screens
