+++ src/Navigation.js
@@ -1,0 +1,19 @@
+import * as React from 'react';
+import {NavigationContainer} from '@react-navigation/native';
+import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
+
+import {HomeStackScreen} from './Stacks/HomeStackScreen';
+import {SettingsStackScreen} from './Stacks/SettingsStackScreen';
+
+const Tab = createBottomTabNavigator();
+
+export default function Navigator() {
+  return (
+    <NavigationContainer>
+      <Tab.Navigator>
+        <Tab.Screen name="Home" component={HomeStackScreen} />
+        <Tab.Screen name="Settings" component={SettingsStackScreen} />
+      </Tab.Navigator>
+    </NavigationContainer>
+  );
+}
