+++ src/Stacks/Screens/DetailsScreen.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import {Text, View} from 'react-native';
+export function DetailsScreen() {
+  return (
+    <View style={{flex: 1, justifyContent: 'center', alignItems: 'center'}}>
+      <Text>Details!</Text>
+    </View>
+  );
+}
