+++ src/Stacks/Screens/HomeScreen.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import {Button, Text, View} from 'react-native';
+export function HomeScreen({navigation}) {
+  return (
+    <View style={{flex: 1, justifyContent: 'center', alignItems: 'center'}}>
+      <Text>Home screen</Text>
+      <Button
+        title="Go to Details"
+        onPress={() => navigation.navigate('Details')}
+      />
+    </View>
+  );
+}
