+++ src/Stacks/Screens/SettingsScreen.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import {Button, Text, View} from 'react-native';
+export function SettingsScreen({navigation}) {
+  return (
+    <View style={{flex: 1, justifyContent: 'center', alignItems: 'center'}}>
+      <Text>Settings screen</Text>
+      <Button
+        title="Go to Details"
+        onPress={() => navigation.navigate('Details')}
+      />
+    </View>
+  );
+}
